Added Shell script equivalent of python script and JSON formatter
Added a script to format JSON formatted items to be a list of JSON
objects.  Takes a file or directory name in which to apply list
formatting on.   Can use wildcards to apply to multiple files at once.
Eventually will be used to insert objects into MongoDB.

gatherNetworkData.sh is a shell script that runs infinitely in a shell
command to store network speed data hourly.
TODO: get it work work as a cronjob.